Extract worker and round logic
